Refactor: character 도멤메인 관련 수정 및 새로운 푸키 ë°발급ê 기능

See merge request s13-webmobile1-sub1/S13P11A604!347

+++ backend/src/main/java/com/ssafy/pookie/character/dto/RepCharacterResponseDto.java
@@ -12,7 +12,7 @@ import lombok.*;
 @Builder
 public class RepCharacterResponseDto {
 
-    private int userCharacterId;
+    private Long userCharacterId;
     private String characterName;
     private int step;
     private int exp;
